pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "azure-storage-blob>=12.27.1",
    "celery>=5.5.3",
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "cryptography>=46.0.3",
    "daphne>=4.2.1",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "factory-boy>=3.3.3",
    "faker>=38.2.0",
    "gunicorn>=23.0.0",
    "msal>=1.34.0",
    "numpy>=2.3.5",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-exporter-otlp>=1.38.0",
    "opentelemetry-instrumentation-django>=0.59b0",
    "opentelemetry-instrumentation-requests>=0.59b0",
    "opentelemetry-sdk>=1.38.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt[crypto]>=2.10.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "whitenoise>=6.11.0",
]
